% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_helper_functions.R
\docType{data}
\name{example_corpus}
\alias{example_corpus}
\title{Demo-Korpus: example_corpus}
\format{
Ein \code{textmeta}-Objekt mit zwei Elementen:
\describe{
\item{text}{Charaktervektor mit Dokumenttexten}
\item{meta}{Dataframe mit Metadaten (id, date, title, resource)}
}
}
\source{
Wikipedia (Texte); Metadaten fiktiv
}
\usage{
example_corpus
}
\description{
Ein Beispielkorpus mit fiktiven Filmtexten für tosca-Analysen.
}
\keyword{datasets}
