% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_helper_functions.R
\name{is.textmeta}
\alias{is.textmeta}
\title{Check whether an object is a textmeta object}
\usage{
is.textmeta(obj)
}
\arguments{
\item{obj}{Any R object.}
}
\value{
Logical scalar: \code{TRUE} if object is a \code{textmeta}, \code{FALSE} otherwise.
}
\description{
Tests whether an object is a valid \pkg{tosca} \code{textmeta} object.
It simply checks if it is a list with \code{meta} and \code{text} components.
}
\examples{
is.textmeta(list(meta = data.frame(), text = list()))
is.textmeta("not a textmeta")
}
