% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_helper_functions.R
\docType{data}
\name{stop_words_de}
\alias{stop_words_de}
\title{stop_word_list: german stop word list}
\format{
Ein \verb{character vector} mit 623 Elementen
}
\source{
https://countwordsfree.com/stopwords
}
\usage{
stop_words_de
}
\description{
Eine Deutsche Stopword-Liste
}
\keyword{datasets}
