% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_helper_functions.R
\docType{data}
\name{stop_words_es}
\alias{stop_words_es}
\title{stop_word_list: spanish stop word list}
\format{
Ein \verb{character vector} mit 608 Elementen
}
\source{
https://countwordsfree.com/stopwords
}
\usage{
stop_words_es
}
\description{
Eine Spanische Stopword-Liste
}
\keyword{datasets}
