% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_helper_functions.R
\docType{data}
\name{stop_words_fr}
\alias{stop_words_fr}
\title{stop_word_list: french stop word list}
\format{
Ein \verb{character vector} mit 558 Elementen
}
\source{
https://countwordsfree.com/stopwords
}
\usage{
stop_words_fr
}
\description{
Eine Franzoesische Stopword-Liste
}
\keyword{datasets}
