% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_helper_functions.R
\docType{data}
\name{stop_words_it}
\alias{stop_words_it}
\title{stop_word_list: italian stop word list}
\format{
Ein \verb{character vector} mit 361 Elementen
}
\source{
https://countwordsfree.com/stopwords
}
\usage{
stop_words_it
}
\description{
Eine Italienische Stopword-Liste
}
\keyword{datasets}
