% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_lda_evaluation.R
\name{topTextsPerUnit}
\alias{topTextsPerUnit}
\title{Export top documents per topic and time period}
\usage{
topTextsPerUnit(
  corpus,
  ldaresult,
  ldaID,
  unit = "quarter",
  nTopTexts = 20,
  tnames = paste0("Topic", 1:K, ".", tosca::topWords(ldaresult$topics)),
  foldername = NULL,
  translate = F,
  max_text_length = 32000,
  source_lang = NULL,
  deepl_key = NULL
)
}
\arguments{
\item{corpus}{A \code{textmeta} object.}

\item{ldaresult}{LDA result from \code{tosca::LDAgen()}.}

\item{ldaID}{Document IDs used in the model.}

\item{unit}{Time unit (\code{"month"}, \code{"bimonth"}, \code{"quarter"}, \code{"halfyear"}, \code{"year"}).}

\item{nTopTexts}{Number of top texts per topic and unit.}

\item{tnames}{Optional custom topic names.}

\item{foldername}{Output folder name. If \code{NULL}, no export is performed.}

\item{translate}{Logical. Translate texts and titles?}

\item{max_text_length}{Max. length of exported text.}

\item{source_lang}{Source language code.}

\item{deepl_key}{DeepL API key.}
}
\value{
Invisibly returns a list of top texts or IDs per topic and time chunk.
}
\description{
Groups documents by time (e.g., month or quarter), then extracts and optionally saves
the most representative texts per topic and time unit.
}
\examples{
# topTextsPerUnit(corpus, ldaresult, ldaID, unit = "quarter", foldername = "toptexts/")
}
